Add pause/resume options
When playing a song, it's useful to have a pause/resume option that
can be triggered by a tag.
For that, there is now a PAUSE and RESUME tag that is handled
specifically.
The implementation uses job control signals (SIGTSTP and SIGCONT) for
the soundkid-player process. A better approach might be using the
librespot API but I was unable to figure out how to do that.